+++ ui/src/main/java/io/snabble/sdk/ui/checkout/CheckoutOnlineView.java
@@ -36,6 +36,7 @@ public class CheckoutOnlineView extends FrameLayout implements Checkout.OnChecko
     private Checkout.State currentState;
     private ImageView helperImage;
     private View upArrow;
+    private View progressIndicator;
 
     public CheckoutOnlineView(Context context) {
         super(context);
@@ -66,6 +67,7 @@ public class CheckoutOnlineView extends FrameLayout implements Checkout.OnChecko
         helperText = findViewById(R.id.helper_text);
         helperImage = findViewById(R.id.helper_image);
         upArrow = findViewById(R.id.arrow);
+        progressIndicator = findViewById(R.id.progress_indicator);
 
         cancel.setOnClickListener(v -> {
             checkout.abort();
@@ -154,6 +156,9 @@ public class CheckoutOnlineView extends FrameLayout implements Checkout.OnChecko
             case PAYMENT_PROCESSING:
                 checkoutIdCode.setVisibility(View.GONE);
                 helperText.setVisibility(View.GONE);
+                helperImage.setVisibility(View.GONE);
+                upArrow.setVisibility(View.GONE);
+                progressIndicator.setVisibility(View.VISIBLE);
                 cancel.setVisibility(View.INVISIBLE);
                 cancelProgress.setVisibility(View.INVISIBLE);
                 break;
@@ -202,10 +207,12 @@ public class CheckoutOnlineView extends FrameLayout implements Checkout.OnChecko
             upArrow.setVisibility(View.VISIBLE);
             helperImage.setVisibility(View.VISIBLE);
             helperText.setVisibility(View.GONE);
+            progressIndicator.setVisibility(View.GONE);
         } else {
             upArrow.setVisibility(View.GONE);
             helperImage.setVisibility(View.GONE);
             helperText.setVisibility(View.VISIBLE);
+            progressIndicator.setVisibility(View.GONE);
         }
     }
 }